Do not wait until IMU data becomes available
IMU data is available at 100hz so nh.spinOnce() was only running
at 100hz as well. This commit removes the waiting for IMU data
so nh.spinOnce() runs at the maximum possible speed reducing
communication latency and making it possible to send messages
at higher rates.

If the motor is stopped the message is send inside the IMU callback
because this runs at 100hz else it would send a twist message at
maximum speed possible. If the motor is running twist messages are
now send whenever we get a motor interrupt.